requirements.txt
chromadb
langchain
google-generativeai
pypdf2
ebooklib
beautifulsoup4
markdown
python-dotenv
langchain_community
langchain-openai
python-docx
email
html2text
tiktoken